+++ idb_shim/lib/src/common/common_meta.dart
@@ -450,12 +450,14 @@ class IdbIndexMeta {
     return metas;
   }
 
-  IdbIndexMeta.fromMap(Map<String, Object?> map) //
-      : this(
-            map['name'] as String?, //
-            map['keyPath'],
-            map['unique'] as bool?, //
-            map['multiEntry'] as bool?);
+  factory IdbIndexMeta.fromMap(Map<String, Object?> map) {
+    var meta = IdbIndexMeta(
+        map['name'] as String?, //
+        map['keyPath'],
+        map['unique'] as bool?, //
+        map['multiEntry'] as bool?);
+    return meta;
+  }
 
   IdbIndexMeta.fromIndex(Index index)
       : this(index.name, index.keyPath, index.unique, index.multiEntry);
@@ -467,7 +469,7 @@ class IdbIndexMeta {
   Map<String, Object?> toMap() {
     dynamic keyPath;
     if (this.keyPath is Iterable) {
-      keyPath = (this.keyPath as Iterable).cast<String>();
+      keyPath = (this.keyPath as Iterable).cast<String>().toList();
     } else {
       keyPath = this.keyPath?.toString();
     }