+++ idb_shim/lib/src/sembast/sembast_object_store.dart
@@ -205,6 +205,14 @@ class ObjectStoreSembast extends ObjectStore with ObjectStoreWithMetaMixin {
 
   @override
   Index createIndex(String name, keyPath, {bool? unique, bool? multiEntry}) {
+    // Be compatible with Chrome
+    if (keyPath is List) {
+      // // Native: InvalidAccessError: Failed to execute 'createIndex' on 'IDBObjectStore': The keyPath argument was an array and the multiEntry option is true.
+      if (multiEntry ?? false) {
+        throw DatabaseError(
+            'The keyPath argument $keyPath cannot be an array if the multiEntry option is true');
+      }
+    }
     final indexMeta = IdbIndexMeta(name, keyPath, unique, multiEntry);
     meta!.createIndex(database.meta, indexMeta);
     return IndexSembast(this, indexMeta);