+++ idb_shim/test/multiplatform/idb_client_memory_test.dart
@@ -32,5 +32,34 @@ void main() {
       await txn.completed;
       db1.close();
     });
+
+    test('index key cursor delete', () async {
+      var factory1 = newIdbFactoryMemory();
+      var db1 =
+          await factory1.open('test', version: 1, onUpgradeNeeded: (event) {
+        final db = event.database;
+        final objectStore =
+            db.createObjectStore(testStoreName, autoIncrement: true);
+        objectStore.createIndex(testNameIndex, testNameField);
+      });
+      var txn = db1.transaction(testStoreName, idbModeReadWrite);
+      var store = txn.objectStore(testStoreName);
+      await store.put({testNameField: 'value1'});
+      var index = store.index(testNameIndex);
+      Object? exception;
+      await index.openKeyCursor(autoAdvance: true).listen((cursor) async {
+        // devPrint('cursor: $cursor');
+        try {
+          await cursor.delete();
+        } catch (e) {
+          exception = e;
+        }
+      }).asFuture();
+
+      await txn.completed;
+
+      db1.close();
+      expect(exception, isA<StateError>());
+    });
   });
 }