+++ idb_shim/tool/run_browser_test.dart
@@ -5,7 +5,7 @@ Future main() async {
 
   await shell.run('''
 
-  pub run test -p chrome
+  dart test -p chrome
 
 ''');
 }