+++ idb_test/lib/index_cursor_test.dart
@@ -21,6 +21,7 @@ void main() {
 }
 
 void defineTests(TestContext ctx) {
+  // final idbFactory = devWarning(getIdbFactoryLogger(ctx.factory));
   final idbFactory = ctx.factory;
   group('index_cursor', () {
     Database? db;
@@ -349,6 +350,36 @@ void defineTests(TestContext ctx) {
               });
         });
       });
+      test('cursor key delete failed', () async {
+        await dbSetUp();
+        dbCreateTransaction();
+        Object? cursorException;
+        var key = add('test1');
+        try {
+          await index.openKeyCursor(autoAdvance: false).listen((Cursor cursor) {
+            // This fails on Chrome, mimic on idb
+            try {
+              cursor.delete().then((_) {
+                cursor.next();
+              });
+            } catch (e) {
+              cursorException = e;
+              print('cursorException: $cursorException');
+              cursor.next();
+            }
+          }).asFuture();
+
+          await transaction!.completed;
+          transaction = db!.transaction(testStoreName, idbModeReadOnly);
+          objectStore = transaction!.objectStore(testStoreName);
+          index = objectStore.index(testNameIndex);
+          var value = await index.get(key);
+          expect(value, isNotNull);
+        } catch (e) {
+          print(e);
+          expect(e, isNot(isA<TestFailure>()));
+        }
+      }, skip: true); // Hangs on Chrome
 
       test('cursor none auto delete 1', () async {
         await dbSetUp();
@@ -849,7 +880,50 @@ void defineTests(TestContext ctx) {
                 direction: idbDirectionPrev)
             .first;
         expect(first.primaryKey, key);
+        var key2 = await objectStore.put({'f1': 1, 'f2': 2, 'f3': 3});
+        expect([key, key2], [1, 2]);
       });
+
+      tearDown(dbTearDown);
+    });
+    group('with_3_keys_unique', () {
+      Future openDb() async {
+        final dbName = ctx.dbName;
+        await idbFactory.deleteDatabase(dbName);
+        void onUpgradeNeeded(VersionChangeEvent e) {
+          final db = e.database;
+          final objectStore =
+              db.createObjectStore(testStoreName, autoIncrement: true);
+          objectStore.createIndex(testNameIndex, ['f1', 'f2', 'f3'],
+              unique: true);
+        }
+
+        db = await idbFactory.open(dbName,
+            version: 1, onUpgradeNeeded: onUpgradeNeeded);
+      }
+
+      tearDown(dbTearDown);
+
+      test('one_record', () async {
+        await openDb();
+        dbCreateTransaction();
+        var key = await objectStore.put({'f1': 1, 'f2': 2, 'f3': 3});
+        expect([key], [1]);
+      }, skip: false); //'open failed'
+      test('two_same_record', () async {
+        await openDb();
+        dbCreateTransaction();
+        var key = await objectStore.put({'f1': 1, 'f2': 2, 'f3': 3});
+        try {
+          await objectStore.put({'f1': 1, 'f2': 2, 'f3': 3});
+          await transaction!.completed;
+          fail('should fail');
+        } on DatabaseError catch (_) {
+          // devPrint('error $e');
+        }
+        transaction = null;
+        expect([key], [1]);
+      }, skip: false); //'open failed');
 
       tearDown(dbTearDown);
     });